MAJ avec multer + ajout et modification d'un thing pour un user

+++ backend/controllers/stuff.js
@@ -2,12 +2,18 @@
 Fichier de contrôleur exporte des méthodes qui sont ensuite attribuées aux routes pour améliorer la maintenabilité de votre application
 */
 
+const { JsonWebTokenError } = require('jsonwebtoken')
 const Thing = require('../models/thing')
+const fs = require("fs")
 
 exports.createThing = (req, res, next) => {
-    delete req.body._id
+    const thingObject = JSON.parse(req.body.thing)
+    delete thingObject._id
+    delete thingObject._userId
     const thing = new Thing({
-        ...req.body
+        ...thingObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
     })
     thing.save()
         .then(() => res.status(201).json({ message: "Objet enregistré !" }))
@@ -17,30 +23,54 @@ exports.createThing = (req, res, next) => {
         })
 }
 
-exports.modifyThing = (req, res, next) =>{
-    Thing.updateOne({_id: req.params.id}, {...req.body, _id: req.params.id})
-    .then(() => res.status(201).json({ message: "Objet enregistré !" }))
-    .catch(error => {
-        console.log(error)
-        res.status(400).json({ error })
+exports.modifyThing = (req, res, next) => {
+    const thingObject = req.file ?{
+        ...JSON.parse(req.body.thing),
+        imageUrl:`${req.protocol}://${req.get("host")}/images/${req.file.filename}`} : {...req.body}
+
+    delete thingObject._userId
+    Thing.findOne({_id: req.params.id})
+    .then(thing =>{
+        if (thing.userId != req.auth.userId) {
+            res.status(401).json({message: "Non AUTORISER"})
+            
+        } else {    
+            Thing.updateOne({ _id: req.params.id }, { ...thingObject, _id: req.params.id })
+                .then(() => res.status(201).json({ message: "Objet enregistré !" }))
+                .catch(error => {res.status(400).json({ error })})
+        }
     })
+    .catch(error => res.status(401).json({error})) 
+
 }
 
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-      .catch(error => res.status(400).json({ error }));
-  }
 
-exports.getOneThing =  (req, res, next) => {
+    Thing.findOne({_id: req.params.id})
+    .then(thing =>{
+        if (thing.userId != req.auth.userId) {
+            res.status(401).json({message: "Non AUTORISER"})            
+        } else {
+            const filename = thing.imageUrl.split("/images")[1]
+            fs.unlink(`images/${filename}`,() =>{
+            Thing.deleteOne({ _id: req.params.id })
+                .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
+                .catch(error => res.status(401).json({ error }));
+            })
+        }
+    })
+    .catch(error => res.status(500).js({error}))
+}
+
+exports.getOneThing = (req, res, next) => {
     Thing.findOne({ _id: req.params.id })
-      .then(thing => res.status(200).json(thing))
-      .catch(error => res.status(404).json({ error }))
-  }
+        .then(thing => res.status(200).json(thing))
+        .catch(error => res.status(404).json({ error }))
+}
 
 exports.getAllThing = (req, res, next) => {
     Thing.find()
-    .then(things => res.status(200).json(things))
-    .catch(error => res.status(400).json({error}))
-   
+        .then(things => res.status(200).json(things))
+        .catch(error => res.status(400).json({ error }))
+
 }
